+++ Source/Core/Core/HW/SI_DeviceGCController.cpp
@@ -289,7 +289,7 @@ void CSIDevice_GCController::SendCommand(u32 _Cmd, u8 _Poll)
     unsigned int uStrength = command.Parameter2;
 
     // get the correct pad number that should rumble locally when using netplay
-    const u8 numPAD = NetPlay_InGamePadToLocalPad(ISIDevice::m_iDeviceNumber);
+    const int numPAD = NetPlay_InGamePadToLocalPad(ISIDevice::m_iDeviceNumber);
 
     if (numPAD < 4)
     {
